Fix the patching in tools/COMPILE-PROTOS.sh for recent protoc versions.
This fixes a problem observed with recent versions of protoc (3.14.0
and possibly 3.13.X). Code in tools/COMPILE-PROTOS.sh "manually"
patches generated code to fix a problem in the automatic CLI generator.

This should be removed when that problem is fixed. We'll need
to file an issue on github.com/googleapis/gapic-generator-go.