Increase portability of proto compilation script
The COMPILE-PROTOS.sh setup script fails when dependencies are installed
to a Go bin directory that isn't accessible by the script (and the
programs it calls). This commit gives the script access to either the
custom or default Go bin directory in case it isn't already accessible.

The script also uses `sed` for string replacement in some files, but the
command behaves differently on the BSD and GNU versions of `sed`. When
the BSD version is used (e.g. on MacOS), then `sed -i -e ...` is
equivalent to `sed -i=-e ...`. When the GNU version is used, then `sed
-i -e ...` considers `-i` and `-e` as separate flags. This commit
updates the `sed` commands to have consistent behavior on BSD and GNU
versions of `sed`.