Add JSON diff for Discovery specs before update
Registry API creates new revisions if the compressed spec content has
changed since the most recent upload. Discovery specs are returned as
JSON and are not guaranteed to be ordered deterministically.
Consequently, the compressed spec may change between update requests
even though the spec is semantically identical.

This change prevents the creation of new spec revisions for semantically
identical discovery specs. Fixing this issue in the Registry API is more
complicated and may be prioritized at a later date.