Refactor + Support revisions for Discovery bulk upload command (#32)
* Refactor Discovery bulk upload command

* Create helpers for common string and processing methods
* Remove special handling of AlreadyExists errors after existence checks
* Defer task queue cleanup after finishing setup
* Use struct literals for construction where possible (readability)
* Decrease variable scopes where possible (readability)
* Make disco script executable

* Support revisions for Discovery bulk upload command

This change adds an attempt to update the spec. The UpdateSpec API
method checks for changes to content and creates a new revision if
necessary.

* Add JSON diff for Discovery specs before update

Registry API creates new revisions if the compressed spec content has
changed since the most recent upload. Discovery specs are returned as
JSON and are not guaranteed to be ordered deterministically.
Consequently, the compressed spec may change between update requests
even though the spec is semantically identical.

This change prevents the creation of new spec revisions for semantically
identical discovery specs. Fixing this issue in the Registry API is more
complicated and may be prioritized at a later date.

* Add missing dependencies to Go module

Co-authored-by: Tim Burks <timburks@google.com>