Create isolated test environment for server (#57)
This change establishes a pattern that can be used for testing all
functionality of the server package. It also restricts the scope of
tests to cover only the server RPC handler functionality.

Tests for the full server stack including gRPC and gapic client behavior
can be found in the `tests/` directory.