Refactor Projects API implementation (#83)
Refactor Projects API implementation

This change is in preparation for incoming functional changes. The
purpose of adding a resource name type is to remove string processing
from our RPC handler implementations as the string manipulation
increases in complexity. Name types will be added to Api/Version/Spec in
a follow-up.

* Add type for project resource names, refactor APIs where applicable.
* Move project storage entity name into storage package.
* Refactor common project operations into helper functions.
* Check additional errors before returning
* Consistently use error helper functions.
* Remove redundant import names.