Core API Test Suite + Bug Fixes (#106)
This change introduces a suite of unit tests for the server. Along with the tests, some bug fixes are included for some situations that tests detected an issue and could be easily resolved. Other known bugs have their tests marked as skipped until they can be addressed on the main branch.