Add logging for GCP project errors
When the storage client cannot be created without error it results in
RPC handlers returning unavailable status. When that happens, apg and
other clients try again until timeout is reached. This change logs the
error when it occurs so developers are informed of their environment
problem. Previously the request would appear to hang.