temporary workaround to fix listing artifacts with the datastore backend
With a recent change, calls to ListArtifacts stopped returning
empty tokens when the end of a collection was reached. That
will be addressd in an upcoming revision, but this temporarily
fixes the problem by using the fact that the next page tokens provided
by the storage backend don't change when the end of the collection
has been reached. When the token that we're returning is the
same as the token we received, we replace the token with an empty string.