Use query offset for pagination
The previous implementation of paginating using keys only works if the
query is ordered by key. Querying for spec revisions ordered by create
time does not necessarily follow the same order as keys, so we would
either have to paginate using create time in this case or generally use
offsets for all types of pagination.

This change introduces offset pagination because it's the more robust
solution and is less complex than determining "safe" fields to use for
pagination depending on the query type.