Minor build fixes (#146)
* add IntelliJ files to .gitignore

* another Mac sed fix

Some versions of BSD sed do not support `\n` as newline in replacement;
you just get an `n`. The replacement text is simple enough to put in a
variable. As a bonus, it is more readable. If it gets more complicated,
will have to do something else.

Also check that the patch was applied.