Reduce timeout to 10s for clients
The previous timeout of 60s doesn't appear to make it more likely to
successfully find and connect to a registry server. Reducing the timeout
to 10s provides quicker feedback that the client is failing to find the
server.

This timeout setting is only a default. In cases where a longer deadline
is desired, clients can be provided a custom deadline using
context.WithDeadline to wrap the existing context.