Add error check to project existence check
This check prevents commands from continuing when they call
EnsureProjectExists but an error occurs while checking the project
existence.

Failing early in this case is preferable because EnsureProjectExists has
failed to ensure that the project exists, and the callers are likely to
rely on this behavior. An error that occurs later due to the project not
existing will increase difficulty of debugging.