Use empty non-nil field masks as implicit updates (#225)
The generated apg tool allows callers to set update_mask.paths with
fields, and defaults to an empty slice if none are provided. Every
request to update a resource from apg has a non-nil mask as a result.

This change adds a check for non-nil but empty field masks and behaves
the same as a nil mask in those cases, i.e. performs an implicit update.