Simplify registry-server configuration.
This commit simplifies server configuration by replacing
Viper-based configuration with a YAML reader that accepts
environment variables in configuration source files. This
allows users to directly see and control the environment
variables that affect configuration and clarifies that
config files have highest priority (making it easier to
debug configuration problems). It also removes the home-
directory-based configuration that seems useful for
command-line tools but adds complexity to server
deployments. The README also clarifies the default
configuration (when no file is specified).

This does not preclude the use of Viper for configuration
of command-line tools like the registry tool, but we should
carefully review pros/cons before committing to that.

This was tested locally with existing CI tests, in Kubernetes
with a ConfigMap-based configuration file, and in Docker with
environment variable configuration.