Replace /contents singleton accessors with custom :getContents methods. (#313)
This is a small breaking change (sorry about that!) that makes the API more consistent with AIP-121 https://google.aip.dev/121). Because "contents" is a field of API specs and artifacts, we decided that it's better to not treat it as a full AIP-121 resource, which has complicating implications for some intended hosting platforms and related tools.