Remove debug logging from gorm package
We should determine how we're handling errors in the gorm package then
introduce consistent and informative logging for each operation.

Currently we only log a silently ignored developer error (incorrect
usage of the Query type) and one database write error while not logging
any other database write errors.

Ideally we can refactor the gorm package interface to return informative
errors that will be logged by its callers who know the severity of the
error (if it was expected, it may only be an info log). The gorm package
would be responsible for informative logging such as SQL queries
executed, connection info, and performance.