Remove flag for custom uid field
This flagging approach interferes with our ability to use context-scoped
loggers because it relies on the PersistentPreRun feature of Cobra. To
use context-scoped loggers, we need to bind a logger to the context
before passing the it to the subcommand constructors that use it.

Because the flag is used to create a "link" between a parent process
executing a command in a child process, we can replace it with other
solutions.

1) We can log the task ID alongside the process ID of the child command,
   and include the process ID of each command in its logs.

2) We can add a configurable value to the environment and include that
   in the logs for each command, setting it uniquely for each child.