Update gorm query to fix ordering bug
Unexpectedly, the Last method in GORM's query builder doesn't select the
last row in the query result. Instead, it adds an additional ordering
parameter to order by descending primary key and takes the first row in
the overall query result. With our queries, it has no effect on the
ordering and results in the first row, i.e. the row with the oldest
revision_create_time, being selected.

This fix just reverses the ordering in our query and selects the first
row, so we always pick the most recent revision_create_time.