Update YAML import/export tests for yaml.v3
The yaml.v3 library uses different indentation than v2,
so the "golden" yaml needed an update to match what was
generated by v3. It also uses 4-character spacing, which
seems excessive for our applications, so this adds a helper
that provides 2-character marshalling.

I also noticed that the comparison order was wrong
in calls to cmp.Diff and fixed that.